Moved all clients to individual modules to avoid bloat
