+++ src/user_interfaces/profile_creation.py
@@ -3,7 +3,8 @@ from typing import List, Tuple
 
 from screeninfo import Monitor
 
-from src.service_clients import VisionTrackingClient, WindowsWebcamClient
+from src.clients.vision_tracking_client import VisionTrackingClient
+from src.clients.windows_webcam_client import WindowsWebcamClient
 from src.user_interfaces.gui_base_class import BaseGUITest
